refactor: extract api and parallax background
